requirements.txt
absl-py==1.3.0
alembic==1.10.2
asttokens==2.2.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.2
Bottleneck==1.3.5
box==0.0.0
brotlipy==0.7.0
cachetools==5.2.0
ccimport==0.4.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.5
cryptography==38.0.1
cumm-cu113==0.4.8
cycler==0.11.0
databricks-cli==0.17.5
debugpy==1.5.1
decorator==5.1.1
distlib==0.3.6
docker==6.0.1
docopt==0.6.2
einops==0.6.0
entrypoints==0.4
executing==1.2.0
filelock==3.10.0
fire==0.5.0
Flask==2.2.3
flit_core==3.6.0
fonttools==4.25.0
fvcore==0.1.5.post20210915
gdown==4.6.4
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.15.0
google-auth-oauthlib==0.4.6
greenlet==2.0.2
grpcio==1.51.1
gunicorn==20.1.0
huggingface-hub==0.13.2
idna==3.4
imageio==2.26.0
importlib-metadata==5.2.0
importlib-resources==5.12.0
iopath==0.1.9
ipykernel==6.15.0
ipython==8.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.1.1
json-tricks==3.16.1
jupyter-client==7.0.6
jupyter_core==4.12.0
kiwisolver==1.4.4
lark==1.1.5
lazy_loader==0.1
llvmlite==0.39.1
lmdb==1.4.0
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.2.2
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mlflow==2.2.2
monai==1.1.0
munkres==1.1.4
nest-asyncio==1.5.6
networkx==3.0
nibabel==5.0.1
ninja==1.11.1
nptyping==2.4.1
numba==0.56.4
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==22.0
pandas==1.4.4
parameterized==0.8.1
parso==0.8.3
pccm==0.4.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==22.3.1
ply==3.11
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.4
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pynrrd==1.0.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-ignite==0.4.11
pytz==2022.7
PyWavelets==1.4.1
PyYAML==5.1.2
pyzmq==19.0.2
querystring-parser==1.2.4
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.1.3
scipy==1.9.1
setproctitle==1.2.2
setuptools==65.5.0
shap==0.41.0
SimpleITK==2.2.1
simplejson==3.18.4
sip==6.6.2
six==1.16.0
slicer==0.0.7
smmap==5.0.0
soupsieve==2.4
spconv-cu113==2.3.5
SQLAlchemy==2.0.6
sqlparse==0.4.3
stack-data==0.6.2
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
termcolor==2.2.0
threadpoolctl==2.2.0
tifffile==2023.3.15
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
torch==1.11.0
torch-tb-profiler==0.4.1
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.27.1
types-pkg-resources==0.1.3
types-PyYAML==6.0.12.8
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.6
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.15.0
yacs==0.1.8
yarg==0.1.9
zipp==3.11.0
